+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [1.1.0] - 2019-03-05
+### Changed
+- Use programmatic debounce for function emitting gpio changes.
+- Do not emit a change event when current value has not changed.
+
 ## [1.0.0] - 2018-12-08
 ### Added
 - First release